Add files via upload
For this project, the data taken from a camera installed in the front part of the car is analyzed, after having been processed to detected and delimited the road lane on the 
images. Two different kind of road models, parabola and clotoide, are also proposed in order to obtain the main variables involved: angular and lateral displacement of the 
vehicle on the road.